Eat all the touchstarts so that we look like a desktop web browser even on mobile and we avoid all this 750ms react-tap-event-plugin nonsense (see the bottom of its readme).
Basically if a page renders to slowly we go over an arbitrary time threshold and click again. Eating the touches fixes that so we only get click events on mobile just like on desktop.

Also move the onboarding and browser check stuff into main's componentDidMount() instead of render().
Now those things only happen when the app is launched. They may not fire if the app was backgrounded, but that's fine for what they are.
(This part will speed up page loads. I noticed it while finding why render was slower in the first place.)

https://github.com/Tour-de-Force/btc-app/issues/160